+++ NRG.CalendarFinder/NRG.CalendarFinder/MsGraphFactories/IMsGraphClientFactory.cs
@@ -4,5 +4,5 @@ namespace NRG.CalendarFinder.MsGraphFactories;
 
 public interface IMsGraphClientFactory
 {
-    public GraphServiceClient GetClient(string name);
+	public GraphServiceClient GetClient(string name);
 }
