+++ NRG.CalendarFinder/NRG.CalendarFinder/MsGraphFactories/Models/MsGraphCredentials.cs
@@ -2,9 +2,9 @@
 
 public record MsGraphCredentials
 {
-    public required string ClientName { get; set; }
-    public required string TenantId { get; set; }
-    public required string ClientId { get; set; }
-    public required string Thumbprint { get; set; }
-    public required string[] Scopes { get; set; }
+	public required string ClientName { get; set; }
+	public required string TenantId { get; set; }
+	public required string ClientId { get; set; }
+	public required string Thumbprint { get; set; }
+	public required string[] Scopes { get; set; }
 }
