+++ NRG.CalendarFinder/NRG.CalendarFinder/Program.cs
@@ -3,7 +3,6 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using NRG.CalendarFinder.Exteinsions;
 using NRG.CalendarFinder.Extensions;
 using NRG.CalendarFinder.Models;
 
@@ -11,46 +10,46 @@ namespace NRG.CalendarFinder;
 
 internal class Program
 {
-    static async Task Main(string[] args)
-    {
-        await Parser.Default.ParseArguments<Options>(args)
-            .WithParsedAsync(RunHost);
-    }
+	static async Task Main(string[] args)
+	{
+		await Parser.Default.ParseArguments<Options>(args)
+			.WithParsedAsync(RunHost);
+	}
 
-    private static async Task RunHost(Options options)
-    {
-        await Console.Out.WriteLineAsync($"Start App.");
+	private static async Task RunHost(Options options)
+	{
+		await Console.Out.WriteLineAsync($"Start App.");
 
-        try
-        {
-            var host = Host.CreateDefaultBuilder()
-                .ConfigureAppConfiguration(builder =>
-                {
-                    builder.AddJsonFile(options.FilePath);
-                })
-                .ConfigureServices((context, services) =>
-                {
-                    // Services 
-                    services.AddSingleton<CalendarFinderService>();
+		try
+		{
+			var host = Host.CreateDefaultBuilder()
+				.ConfigureAppConfiguration(builder =>
+				{
+					builder.AddJsonFile(options.FilePath);
+				})
+				.ConfigureServices((context, services) =>
+				{
+					// Services 
+					services.AddSingleton<CalendarFinderService>();
 
-                    // Workers
-                    services.AddHostedService<CalendarFinderWorker>();
-                })
-                .AddProcessDataFromJson(options)
-                .AddGraphClientsFromJson()
-                .UseConsoleLifetime()
-                .ConfigureLogging(e => e.SetMinimumLevel(LogLevel.None))
-                .Build();
+					// Workers
+					services.AddHostedService<CalendarFinderWorker>();
+				})
+				.AddProcessDataFromJson(options)
+				.AddGraphClientsFromJson()
+				.UseConsoleLifetime()
+				.ConfigureLogging(e => e.SetMinimumLevel(LogLevel.None))
+				.Build();
 
-            await host.RunAsync();
-        }
-        catch (Exception ex)
-        {
-            await Console.Out.WriteLineAsync($"Failed with error: {ex.Message}");
-        }
-        finally
-        {
-            await Console.Out.WriteLineAsync($"Terminate App.");
-        }
-    }
+			await host.RunAsync();
+		}
+		catch (Exception ex)
+		{
+			await Console.Out.WriteLineAsync($"Failed with error: {ex.Message}");
+		}
+		finally
+		{
+			await Console.Out.WriteLineAsync($"Terminate App.");
+		}
+	}
 }
